requirements.txt
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1 
flask-cors==4.0.0
psycopg2-binary==2.9.9
Flask-Migrate== 4.0.5 
python-dotenv == 1.0.0